Move `get_sector_from_single_sector_obs_id` to TessDataInterface
