+++ infer.py
@@ -30,6 +30,6 @@ for example_path in example_paths:
     label, prediction = database.inference_postprocessing(label, prediction, times.shape[0])
     tess_data_interface = TessDataInterface()
     tic_id = tess_data_interface.get_tic_id_from_single_sector_obs_id(Path(example_path).stem)
-    sector = database.get_sector_from_single_sector_obs_id(Path(example_path).stem)
+    sector = tess_data_interface.get_sector_from_single_sector_obs_id(Path(example_path).stem)
     plot_title = f'TIC {tic_id} sector {sector}'
     plot_lightcurve(times, fluxes, label, prediction, title=plot_title, save_path=f'{plot_title}.png')
