+++ ramjet/photometric_database/toi_lightcurve_database.py
@@ -71,9 +71,8 @@ class ToiLightcurveDatabase(TessTransitLightcurveLabelPerTimeStepDatabase):
                                                      dispositions['transit_duration'].notna()]
         lightcurve_paths = list(self.lightcurve_directory.glob('*lc.fits'))
         tess_data_interface = TessDataInterface()
-        tic_ids = [int(tess_data_interface.get_tic_id_from_single_sector_obs_id(path.name))
-                   for path in lightcurve_paths]
-        sectors = [self.get_sector_from_single_sector_obs_id(path.name) for path in lightcurve_paths]
+        tic_ids = [tess_data_interface.get_tic_id_from_single_sector_obs_id(path.name) for path in lightcurve_paths]
+        sectors = [tess_data_interface.get_sector_from_single_sector_obs_id(path.name) for path in lightcurve_paths]
         lightcurve_meta_data = pd.DataFrame({'lightcurve_path': list(map(str, lightcurve_paths)), 'tic_id': tic_ids,
                                              'sector': sectors})
         meta_data_frame_with_candidate_nans = pd.merge(confirmed_planet_dispositions, lightcurve_meta_data,
